added a testcase for install order
this tests whether the order of plugins, groups as well as the priority,
alwaysInstall and installIfUsable are all obeyed correctly.

The way this works is that the mod installs a bunch of text files and if all
ordering features are obeyed correctly all text files should be empty. If the
feature doesn't work correctly the file will have content describing what
(probably) doesn't work, although the tests for some features (e.g. alwaysinstall
and installifusable) depend on others (e.g. grouporder) to work.