fixed permissions when run in a container
previously we tried to allow access to the directory containing the installer
which may not be allowed in production builds.
Now the script types are loaded statically (not like anyone is writing plugins
for this) and thus we should get away with allowing access only to the
individual files - which seems to work.