removed the code around granting access to the installer exe/dir
From now on it's up to the application to grant the necessary rights to
that directory (specifically: giving read/execute rights to the special
user accounts "ALL APPLICATION PACKAGES"/"ALL RESTRICTED APPLICATION PACKAGES"
aka "S-1-15-2-1" and "S-1-15-2-2".

This change also implements a mode for starting the installer process as
a low-integrity process but that path currently doesn't work because we
can't open an outgoing pipe in that integrity mode, at least not with default
options and working around this (e.g. by using stdin/stdout to communicate
with the host) would require massive changes to both sides.
I'm leaving this in for now in case we come up with a better solution for the
communication problem.